Added models.py Created flask factory and configured bootstrap, ckeditor.
