+++ app/__init__.py
@@ -1,5 +1,7 @@
 # make our app a package
 from flask import Flask
+from flask_bootstrap import Bootstrap
+from flask_ckeditor import CKEditor
 from flask_sqlalchemy import SQLAlchemy
 
 db = SQLAlchemy()
@@ -15,14 +17,14 @@ def create_app():
         app.config.from_object("config.Config")
 
     db.init_app(app)
-
+    ckeditor = CKEditor(app)
+    Bootstrap(app)
     with app.app_context():
         from . import user_views, public_views  # Import routes
         db.create_all()  # Create sql tables for our data models
 
         return app
 
+
 # importing views files to avoid a circular import
-from app import public_views
-from app import admin_views  # import admin views
-from app import user_views
+#
