+++ app/public_views.py
@@ -1,9 +1,9 @@
-from app import app
+from flask import current_app as app
 from werkzeug.security import generate_password_hash, check_password_hash
 from flask_login import login_user, current_user
 from flask import render_template, redirect, url_for, flash
 from app.forms import RegisterForm, LoginForm
-from main import User
+from .models import db, User
 
 
 @app.route('/register', methods=["GET", "POST"])
